feat: dayspac supports float fix: rm qzemoji.init fix: discriminate emoji and normal text

+++ src/aioqzone_feed/api/feed.py
@@ -104,7 +104,7 @@ class FeedApi(Emittable[FeedEvent]):
 
     async def get_feeds_by_second(
         self,
-        seconds: int,
+        seconds: float,
         start: Optional[float] = None,
         *,
         exceed_pred: Optional[Callable[[FeedRep], Awaitable[bool]]] = None,
